+++ backend/app/Http/Requests/Auth/ForgotPasswordRequest.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Http\Requests\Requests\Auth;
+namespace App\Http\Requests\Auth;
 
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class VerifyOtpRequest extends FormRequest
+class ForgotPasswordRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
      */
     public function authorize(): bool
     {
-        return true;
+        return false;
     }
 
     /**
@@ -23,7 +23,7 @@ class VerifyOtpRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'otp' => 'required|digits:6|exists:password_reset_otps,otp',
+            'email' => 'required|email|exists:users,email'
         ];
     }
 }
