+++ backend/app/Http/Requests/Auth/ResetPasswordRequest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Requests\Requests\Auth;
+namespace App\Http\Requests\Auth;
 
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class RegisterRequest extends FormRequest
+class ResetPasswordRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -23,10 +23,8 @@ class RegisterRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'name' => 'required|string|max:255',
-            'email' => 'required|email|unique:users,email',
+            'otp' => 'required|exists:password_reset_otps,otp',
             'password' => 'required|string|min:8',
-            'password_confirm' => 'required|string|same:password',
         ];
     }
 }
